AT MODIFIED worked on fixing bugs with Add to cart Also did some directory neatening and naming convention edits assets/scripts/auth/api.js - removed unnecessary require - removed product functions (addProduct, emptyCart) and moved them to scripts/products/ assets/scripts/auth/events.js - removed unnecessary require - removed product functions (onAddProduct, onEmpty Cart) and moved them to scripts/products/ assets/scripts/auth/ui.js - removed product functions (onAddProductSuccess/Failure, onEmptyCartSuccess/Failure) and moved them to scripts/products/ assets/scripts/index.js - changed const produce to const products for more consistent naming conventions - add jquery to hide the "empty cart" button until logged in - changed Add to cart jquery to simplify assets/scripts/products/api.js - formerly produce/api.js - changed #productList .append to include product id as the add-to-cart button id assets/scripts/products/events.js - formerly produce/events.js - got product id and user id to show in onAddProduct assets/scripts/products/ui.js - formerly produce/ui.js assets/styles/formatting.scss - changed <h4> color to be consistent with color scheme - changed naming convention from #cartBtn to #add-to-cart-btn for index.html - deleted font scripts TODO - changed format of #emptyCart to empty-cart for consistency - changed format of shopBtn to #show-my-cart for consistency
