+++ assets/scripts/auth/api.js
@@ -1,6 +1,5 @@
 'use strict'
 const app = require('../app.js')
-const config = require('../config.js')
 
 // POST (signup)
 const signUp = function (data) {
@@ -67,40 +66,9 @@ const changePassword = (data) => {
   })
 }
 
-// Add Product to user Cart
-const addProduct = (title, price, data) => {
-  return $.ajax({
-    url: app.host + '/users/' + app.user.id,
-    method: 'PATCH',
-    headers: {
-      Authorization: 'Token token=' + app.user.token
-    },
-    data: {
-      'products': {
-        'title': title,
-        'price': price
-      }
-    }
-  })
-}
-
-const emptyCart = (data) => {
-  console.log(data)
-  return $.ajax({
-    url: app.host + '/users/' + app.user.id + '/cart',
-    method: 'PATCH',
-    headers: {
-      Authorization: 'Token token=' + app.user.token
-    },
-    data: app.user.cart
-  })
-}
-
 module.exports = {
   signUp,
   signIn,
   signOut,
-  changePassword,
-  addProduct,
-  emptyCart
+  changePassword
 }
