+++ assets/scripts/auth/events.js
@@ -2,7 +2,6 @@
 const api = require('./api.js')
 const getFormFields = require('../../../lib/get-form-fields.js')
 const ui = require('./ui.js')
-const app = require('../app.js')
 
 // POST (ash - signup)
 const onSignUp = function (event) {
@@ -45,52 +44,9 @@ const onChangePassword = function (event) {
     .fail(ui.changePasswordFail)
 }
 
-const onAddProduct = function (event) {
-  console.log('Step 1: Events Start')
-  event.preventDefault()
-  const pp = function () {
-    const productCollection = $('#productList').children()
-    const priceLocator = productCollection.children().first().siblings().next().html()
-    const prodTypeMod = priceLocator.split('').splice(2, 6).join('')
-    const prodPrice = parseInt(prodTypeMod) + 0.99
-    return prodPrice
-  }
-  const price = pp()
-  console.log(price)
-  console.log('Got Here, got price')
-  const pt = function () {
-    const productCollection = $('#productList').children()
-    const prodTitle = productCollection.children().first().children().html()
-    return prodTitle
-  }
-  const title = pt()
-  console.log('got here 3')
-  console.log(title)
-  let user = app.host.user
-  console.log('Events User')
-  console.log(user)
-  api.addProduct(title, price)
-    .then(ui.onAddProductSuccess)
-    .catch(ui.onAddProductFailure)
-}
-
-const onEmptyCart = function (event) {
-  console.log('Step 1: Events Start')
-  // event.preventDefault()
-  // const data = event
-  console.log('data')
-  // console.log(data)
-  api.emptyCart()
-    .then(ui.onAddProductSuccess)
-    .catch(ui.onAddProductFailure)
-}
-
 module.exports = {
   onSignUp,
   onSignIn,
   onSignOut,
-  onChangePassword,
-  onAddProduct,
-  onEmptyCart
-
+  onChangePassword
 }
