+++ assets/scripts/auth/ui.js
@@ -1,8 +1,6 @@
 'use strict'
 const app = require('../app.js')
 
-
-
 // POST (signup)
 const signUpSuccess = (data) => {
   app.user = data.user
@@ -30,6 +28,7 @@ const signInSuccess = (data) => {
   $('.user-signout').show()
   $('#show-change-pw').show()
   $('#show-my-cart').show()
+  $('#empty-cart').hide()
 }
 
 const signInFail = (error) => {
@@ -78,31 +77,6 @@ const changePasswordFail = (error) => {
   $('#show-change-pw').show()
 }
 
-const onAddProductSuccess = (data) => {
-  console.log('success')
-  console.log(app.user.cart)
-  console.log(app.user.cart.length)
-
-  for (let i = 0; i < app.user.cart.length; i++) {
-    $('#cartContainer').append(
-      '<li>' + app.user.cart[i].title + ' $ ' + app.user.cart[i].price + '</li>'
-    )
-  }
-}
-
-const onAddProductFailure = (data) => {
-  console.log('failure')
-}
-
-const onEmptyCartSuccess = (data) => {
-  $('#cartContainer').remove('li')
-  console.log('should be empty')
-}
-
-const onEmptyCartFailure = (data) => {
-  console.log('shit')
-}
-
 module.exports = {
   signUpSuccess,
   signUpFail,
@@ -111,9 +85,5 @@ module.exports = {
   signOutSuccess,
   signOutFail,
   changePasswordSuccess,
-  changePasswordFail,
-  onAddProductSuccess,
-  onAddProductFailure,
-  onEmptyCartFailure,
-  onEmptyCartSuccess
+  changePasswordFail
 }
