+++ assets/scripts/index.js
@@ -3,7 +3,8 @@
 const setAPIOrigin = require('../../lib/set-api-origin')
 const config = require('./config')
 const authEvents = require('./auth/events.js')
-const produce = require('./produce/api.js')
+const productEvents = require('./products/events.js')
+const products = require('./products/api.js')
 
 $(() => {
   setAPIOrigin(location, config)
@@ -14,44 +15,42 @@ $(() => {
   $('#show-change-pw').hide()
   $('.change-password').hide()
   $('#show-my-cart').hide()
-})
+  $('#empty-cart').hide()
 
-$('.productBtn').on('load', produce.getProduct())
+  $('.productBtn').on('load', products.getProduct())
 
-$('#show-signup').click(function () {
-  $('.user-signup').show()
-  $('#show-signin').hide()
-  $('#show-signup').hide()
-  $('#already-prompt').hide()
-})
+  $('#show-signup').click(function () {
+    $('.user-signup').show()
+    $('#show-signin').hide()
+    $('#show-signup').hide()
+    $('#already-prompt').hide()
+  })
 
-$('#show-signin').click(function () {
-  $('.user-signin').show()
-  $('#show-signin').hide()
-  $('#show-signup').hide()
-  $('#already-prompt').hide()
-})
+  $('#show-signin').click(function () {
+    $('.user-signin').show()
+    $('#show-signin').hide()
+    $('#show-signup').hide()
+    $('#already-prompt').hide()
+  })
 
-$('#show-change-pw').click(function () {
-  $('.change-password').show()
-  $('#show-change-pw').hide()
-  $('#show-my-cart').hide()
-})
+  $('#show-change-pw').click(function () {
+    $('.change-password').show()
+    $('#show-change-pw').hide()
+    $('#show-my-cart').hide()
+  })
 
-$('.user-signup').on('submit', authEvents.onSignUp)
-$('.user-signin').on('submit', authEvents.onSignIn)
-$('.user-signout').on('submit', authEvents.onSignOut)
-// $('#show-change-pw')
-$('.change-password').on('submit', authEvents.onChangePassword)
+  $('.user-signup').on('submit', authEvents.onSignUp)
+  $('.user-signin').on('submit', authEvents.onSignIn)
+  $('.user-signout').on('submit', authEvents.onSignOut)
+  // $('#show-change-pw')
+  $('.change-password').on('submit', authEvents.onChangePassword)
 
-$('#productList').click(function () {
-  console.log('wtf')
-  $('#productList').click(authEvents.onAddProduct)
-})
+  $(document).on('click', '.add-to-cart-btn', productEvents.onAddProduct)
 
-const empty = authEvents.onEmptyCart
-$('#emptyCart').click(function () {
-  empty()
+  const empty = authEvents.onEmptyCart
+  $('#empty-cart').click(function () {
+    empty()
+  })
 })
 
 // const pp = function () {
