+++ assets/styles/typography.scss
@@ -1,7 +1,4 @@
 $header-font: 'Berkshire Swash', cursive;
-
 $body-font: 'News Cycle', sans-serif;
-
 $tagline-font: 'Sacramento', cursive;
-
 $sub-font: 'PT Sans Narrow', sans-serif;
